requirements.txt
fastapi
redis
faker
strawberry-graphql[fastapi]
uvicorn[standard]
broadcaster
peewee